fix: check if file exists before removing

+++ tools/fold_generator.py
@@ -52,8 +52,10 @@ for i in range(1, num_folds + 1):
     trainfile = input_dir + '/train' + str(i) + '.txt'
     if os.path.exists(trainfile):
         os.remove(trainfile)
-os.remove(input_dir + '/test.txt')
-os.remove(input_dir + '/train.txt')
+if os.path.isfile(input_dir + '/test.txt')
+    os.remove(input_dir + '/test.txt')
+if os.path.isfile(input_dir + '/train.txt')
+    os.remove(input_dir + '/train.txt')
 
 # maintain class index
 class_index = 0
